SelectMenu/getPrev+getNextIndex: 87.5%/97.4% -> 100%/100%
- Return type s8 (not int): callee skips return-path extsb since
  EABI delegates s8 sign-extension to the caller. Pairs with `s8 result`
  to give the target's `extsb r5, r4` in-loop AND `mr r3, r5` at tail.
- getPrev: single-var loop `for (s32 i = idx - 1; i >= 0; i--)` lets
  MWCC set CR0 on the subtraction (subic.) and use blt against signed
  decrement, matching target's subic./addi/blt pattern (vs our prior
  subi/addic./ble from separate loop counter).
- The MWCC.md "trade-off" note about s8-result emitting extsb-at-return
  was a misread: target uses mr at return because the function actually
  returns s8, not int.

+++ include/GC2D/SelectMenu.hpp
@@ -30,8 +30,8 @@ public:
 	              TSelectDir* dir);
 	void startMove();
 	void startOpenWindow();
-	int getNextIndex();
-	int getPrevIndex();
+	s8 getNextIndex();
+	s8 getPrevIndex();
 
 public:
 	/* 0x10 */ u32 mState;
